feat: add show dto and update security movie service for delete movie

+++ src/main/java/com/smitsatwara/cinebook/service/MovieService.java
@@ -36,6 +36,9 @@ public class MovieService {
     }
 
     public void deleteMovie(Long movieId) {
+        if(!movieRepository.existsById(movieId)){
+            throw new RuntimeException("Movie not found with id: " + movieId);
+        }
         movieRepository.deleteById(movieId);
     }
